broker: eliminate publisher class
Problem: all that remains of the broker publisher class
is a couple of message handlers for subscription management.

Fold that back into broker.c and eliminate the publisher class.
